Implement link style and action
